Restored the DBF table import (signature of implemented function did not match the one declared in parent class)
git-svn-id: https://svn.52north.org/svn/ilwis/drawers@2734 b64c5c0c-1e74-440b-9268-4ad09cf26390

+++ Sources/Engine/DataExchange/TableDBF.cpp
@@ -400,7 +400,7 @@ void TableDBF::ProcessHeader(File &InputFile)
 	InputFile.Seek(hdrDBF.iHeaderSize);
 }
 
-bool TableDBF::SplitLine(File &InputFile, FieldValues& line, int& iRec)
+bool TableDBF::SplitLine(File &InputFile, FieldValues& line, long& iRec)
 {
 	if ( InputFile.cReadChar() != ' ') // delete records are skipped
 		return false;
@@ -444,7 +444,7 @@ long TableDBF::iNewRecord(long iRec)
 	return TableExternalFormat::iNewRecord(iRec);
 }
 
-void TableDBF::PutExternalField(FieldValues &line, int &iMemoCol, int iRec, int iField)
+void TableDBF::PutExternalField(FieldValues &line, int &iMemoCol, long iRec, int iField)
 {
 	memos[iMemoCol].iValues[iRec - 1 ] = line[iField].iVal();
 //	int iC = memos[iMemoCol][iRec - 1];
